# docs/swagger.yaml
basePath: /
definitions:
  internal_api.AccountResponse:
    properties:
      balance:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      owner_id:
        description: nullable
        type: string
    type: object
  internal_api.EntryResponse:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      credit:
        type: string
      debit:
        type: string
      description:
        type: string
      id:
        type: string
      operation_type:
        type: string
      transaction_id:
        type: string
    type: object
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_api.MessageResponse:
    properties:
      message:
        type: string
    type: object
  internal_api.RegisterResponse:
    properties:
      email:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  internal_api.TokenResponse:
    properties:
      token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Production-grade double-entry accounting ledger
  title: Double-Entry Bank Ledger API
  version: "1.0"
paths:
  /accounts:
    get:
      description: Returns list of accounts owned by authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.AccountResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: List user accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Creates a new user-owned account with name and currency
      parameters:
      - description: Account details
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      description: Returns details of a specific account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get account details
      tags:
      - accounts
  /accounts/{id}/deposit:
    post:
      consumes:
      - application/json
      description: Deposits fiat amount (mock) with double-entry ledger update
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Deposit amount (e.g., 1000.0000)
        in: body
        name: body
        required: true
        schema:
          properties:
            amount:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Deposit money into account
      tags:
      - accounts
  /accounts/{id}/entries:
    get:
      description: Returns list of ledger entries for an account (immutable history)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api.EntryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Get account entries
      tags:
      - accounts
  /accounts/{id}/reconcile:
    get:
      description: Verifies stored balance matches sum of all ledger entries (credits
        - debits)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              matched:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Reconcile account balance
      tags:
      - accounts
  /accounts/{id}/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraws fiat amount (mock) with double-entry ledger update
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Withdraw amount (e.g., 500.0000)
        in: body
        name: body
        required: true
        schema:
          properties:
            amount:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Withdraw money from account
      tags:
      - accounts
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates user with email/password and returns JWT token
      parameters:
      - description: User login details
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user with email and hashed password, returns user
        details and JWT token
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /transfers:
    post:
      consumes:
      - application/json
      description: Transfers fiat amount (mock) with double-entry ledger update
      parameters:
      - description: Transfer details
        in: body
        name: body
        required: true
        schema:
          properties:
            amount:
              type: string
            from_id:
              type: string
            to_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - Bearer: []
      summary: Transfer money between accounts
      tags:
      - accounts
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"